Extract UTs
