+++ analyzers/tests/SonarAnalyzer.UnitTest/SymbolicExecution/Roslyn/RoslynSymbolicExecutionTest.Branching.LearnConstraint.cs
@@ -1,0 +1,104 @@
+﻿/*
+ * SonarAnalyzer for .NET
+ * Copyright (C) 2015-2022 SonarSource SA
+ * mailto: contact AT sonarsource DOT com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+
+using SonarAnalyzer.SymbolicExecution.Constraints;
+using SonarAnalyzer.SymbolicExecution.Roslyn;
+using SonarAnalyzer.UnitTest.TestFramework.SymbolicExecution;
+
+namespace SonarAnalyzer.UnitTest.SymbolicExecution.Roslyn
+{
+    public partial class RoslynSymbolicExecutionTest
+    {
+        [TestMethod]
+        public void Branching_BoolSymbol_LearnsBoolConstraint()
+        {
+            const string code = @"
+if (boolParameter)          // True constraint is learned
+{
+    Tag(""True"", boolParameter);
+    if (boolParameter)      // True constraint is known
+    {
+        Tag(""TrueTrue"");
+    }
+    else
+    {
+        Tag(""TrueFalse Unreachable"");
+    }
+}
+else                        // False constraint is learned
+{
+    Tag(""False"", boolParameter);
+    if (boolParameter)      // False constraint is known
+    {
+        Tag(""FalseTrue Unreachable"");
+    }
+    else
+    {
+        Tag(""FalseFalse"");
+    }
+};
+Tag(""End"");";
+            var validator = SETestContext.CreateCS(code).Validator;
+            validator.ValidateTagOrder(
+                "True",
+                "False",
+                "TrueTrue",
+                "FalseFalse",
+                "End");
+            validator.ValidateTag("True", x => x.HasConstraint(BoolConstraint.True).Should().BeTrue());
+            validator.ValidateTag("False", x => x.HasConstraint(BoolConstraint.False).Should().BeTrue());
+        }
+
+        [TestMethod]
+        public void Branching_BoolOperation_LearnsBoolConstraint()
+        {
+            const string code = @"
+if (collection.IsReadOnly)
+{
+    Tag(""If"", collection);
+}
+Tag(""End"", collection);";
+            var check = new ConditionEvaluatedTestCheck(x => x.State[x.Operation].HasConstraint(BoolConstraint.True)
+                                                                 ? x.SetSymbolConstraint(x.Operation.Instance.AsPropertyReference().Value.Instance.TrackedSymbol(), DummyConstraint.Dummy)
+                                                                 : x.State);
+            var validator = SETestContext.CreateCS(code, ", ICollection<object> collection", check).Validator;
+            validator.ValidateTag("If", x => x.HasConstraint(DummyConstraint.Dummy).Should().BeTrue());
+            validator.TagStates("End").Should().HaveCount(2);
+        }
+
+        [TestMethod]
+        public void Branching_BoolExpression_LearnsBoolConstraint_NotSupported()
+        {
+            const string code = @"
+if (boolParameter == true)
+{
+    Tag(""BoolParameter"", boolParameter);
+}
+bool value;
+if (value = boolParameter)
+{
+    Tag(""Value"", value);
+}";
+            var validator = SETestContext.CreateCS(code).Validator;
+            validator.ValidateTag("BoolParameter", x => x.Should().BeNull());
+            validator.ValidateTag("Value", x => x.Should().BeNull());
+        }
+    }
+}
