+++ analyzers/tests/SonarAnalyzer.UnitTest/SymbolicExecution/Roslyn/RoslynSymbolicExecutionTest.Branching.cs
@@ -485,81 +485,6 @@ Tag(""End"");";
         }
 
         [TestMethod]
-        public void Branching_BoolSymbol_LearnsBoolConstraint()
-        {
-            const string code = @"
-if (boolParameter)          // True constraint is learned
-{
-    Tag(""True"", boolParameter);
-    if (boolParameter)      // True constraint is known
-    {
-        Tag(""TrueTrue"");
-    }
-    else
-    {
-        Tag(""TrueFalse Unreachable"");
-    }
-}
-else                        // False constraint is learned
-{
-    Tag(""False"", boolParameter);
-    if (boolParameter)      // False constraint is known
-    {
-        Tag(""FalseTrue Unreachable"");
-    }
-    else
-    {
-        Tag(""FalseFalse"");
-    }
-};
-Tag(""End"");";
-            var validator = SETestContext.CreateCS(code).Validator;
-            validator.ValidateTagOrder(
-                "True",
-                "False",
-                "TrueTrue",
-                "FalseFalse",
-                "End");
-            validator.ValidateTag("True", x => x.HasConstraint(BoolConstraint.True).Should().BeTrue());
-            validator.ValidateTag("False", x => x.HasConstraint(BoolConstraint.False).Should().BeTrue());
-        }
-
-        [TestMethod]
-        public void Branching_BoolOperation_LearnsBoolConstraint()
-        {
-            const string code = @"
-if (collection.IsReadOnly)
-{
-    Tag(""If"", collection);
-}
-Tag(""End"", collection);";
-            var check = new ConditionEvaluatedTestCheck(x => x.State[x.Operation].HasConstraint(BoolConstraint.True)
-                                                                 ? x.SetSymbolConstraint(x.Operation.Instance.AsPropertyReference().Value.Instance.TrackedSymbol(), DummyConstraint.Dummy)
-                                                                 : x.State);
-            var validator = SETestContext.CreateCS(code, ", ICollection<object> collection", check).Validator;
-            validator.ValidateTag("If", x => x.HasConstraint(DummyConstraint.Dummy).Should().BeTrue());
-            validator.TagStates("End").Should().HaveCount(2);
-        }
-
-        [TestMethod]
-        public void Branching_BoolExpression_LearnsBoolConstraint_NotSupported()
-        {
-            const string code = @"
-if (boolParameter == true)
-{
-    Tag(""BoolParameter"", boolParameter);
-}
-bool value;
-if (value = boolParameter)
-{
-    Tag(""Value"", value);
-}";
-            var validator = SETestContext.CreateCS(code).Validator;
-            validator.ValidateTag("BoolParameter", x => x.Should().BeNull());
-            validator.ValidateTag("Value", x => x.Should().BeNull());
-        }
-
-        [TestMethod]
         public void Branching_ConditionEvaluated()
         {
             const string code = @"
